refactor(HVROverrides): Moved HVR overrides into separate namespace for better visibility.
